Address CR feedback
Remove highlevel.common.ts.  The interfaces are moved into clients.
