+++ sdk/storage/storage-blob/src/BlobClient.ts
@@ -16,7 +16,6 @@ import { setURLParameter } from "./utils/utils.common";
 import { AppendBlobClient } from "./internal";
 import { BlockBlobClient } from "./internal";
 import { PageBlobClient } from "./internal";
-import { DownloadFromBlobOptions } from "./highlevel.common";
 import { Batch } from "./utils/Batch";
 import { streamToBuffer } from "./utils/utils.node";
 
@@ -455,6 +454,75 @@ export interface BlobSetTierOptions {
    * @memberof BlobSetTierOptions
    */
   leaseAccessConditions?: Models.LeaseAccessConditions;
+}
+
+/**
+ * Option interface for BlobClient.downloadToBuffer().
+ *
+ * @export
+ * @interface DownloadFromBlobOptions
+ */
+export interface DownloadFromBlobOptions {
+  /**
+   * Aborter instance to cancel request. It can be created with Aborter.none
+   * or Aborter.timeout(). Go to documents of {@link Aborter} for more examples
+   * about request cancellation.
+   *
+   * @type {Aborter}
+   * @memberof IUploadToBlockBlobOptions
+   */
+  abortSignal?: Aborter;
+
+  /**
+   * blockSize is the data every request trying to download.
+   * Must be >= 0, if set to 0 or undefined, blockSize will automatically calculated according
+   * to the blob size.
+   *
+   * @type {number}
+   * @memberof DownloadFromBlobOptions
+   */
+  blockSize?: number;
+
+  /**
+   * Optional. ONLY AVAILABLE IN NODE.JS.
+   *
+   * How many retries will perform when original block download stream unexpected ends.
+   * Above kind of ends will not trigger retry policy defined in a pipeline,
+   * because they doesn't emit network errors.
+   *
+   * With this option, every additional retry means an additional FileClient.download() request will be made
+   * from the broken point, until the requested block has been successfully downloaded or
+   * maxRetryRequestsPerBlock is reached.
+   *
+   * Default value is 5, please set a larger value when in poor network.
+   *
+   * @type {number}
+   * @memberof DownloadFromAzureFileOptions
+   */
+  maxRetryRequestsPerBlock?: number;
+
+  /**
+   * Progress updater.
+   *
+   * @memberof DownloadFromBlobOptions
+   */
+  progress?: (progress: TransferProgressEvent) => void;
+
+  /**
+   * Access conditions headers.
+   *
+   * @type {BlobAccessConditions}
+   * @memberof DownloadFromBlobOptions
+   */
+  blobAccessConditions?: BlobAccessConditions;
+
+  /**
+   * Concurrency of parallel download.
+   *
+   * @type {number}
+   * @memberof DownloadFromBlobOptions
+   */
+  parallelism?: number;
 }
 
 /**
