+++ sdk/storage/storage-file/src/index.browser.ts
@@ -11,7 +11,6 @@ export * from "./DirectoryClient";
 export * from "./FileClient";
 export * from "./credentials/AnonymousCredential";
 export * from "./credentials/Credential";
-export * from "./highlevel.common";
 export { IPRange as IPRange } from "./IPRange";
 export { Range } from "./Range";
 export * from "./Pipeline";
